Plan 202: split cmd/mdsmith/main.go into per-subcommand files
Move runCheck/checkStdin/checkFiles/checkDiscovered into check.go,
runFix/parseFixFlags/fixCLIOpts/fix*/reportFix*/dry-run helpers into
fix.go, and the runHelp dispatch + topic handlers into help.go. Mirror
the parseFixFlags pattern for check: extract checkCLIOpts + a
parseCheckFlags helper so runCheck is a thin dispatcher, and extract
reportCheckResult so checkStdin/checkFiles/checkDiscovered no longer
duplicate the diagnostic/stats/exit-code block. main.go drops from
1542 to 665 lines; every handler body is under ~50 lines.
