+++ cmd/mdsmith/main.go
@@ -1,31 +1,22 @@
 package main
 
 import (
-	"encoding/json"
 	"errors"
 	"fmt"
 	"io"
-	"math"
 	"os"
 	"path/filepath"
 	"runtime/debug"
-	"sort"
-	"strings"
 
 	flag "github.com/spf13/pflag"
 
-	"github.com/jeduden/mdsmith/internal/concepts"
 	"github.com/jeduden/mdsmith/internal/config"
 	"github.com/jeduden/mdsmith/internal/discovery"
-	"github.com/jeduden/mdsmith/internal/engine"
-	fixpkg "github.com/jeduden/mdsmith/internal/fix"
 	"github.com/jeduden/mdsmith/internal/lint"
 	vlog "github.com/jeduden/mdsmith/internal/log"
 	"github.com/jeduden/mdsmith/internal/output"
 	"github.com/jeduden/mdsmith/internal/profiling"
 	"github.com/jeduden/mdsmith/internal/query"
-	"github.com/jeduden/mdsmith/internal/rule"
-	ruledocs "github.com/jeduden/mdsmith/internal/rules"
 	"github.com/jeduden/mdsmith/internal/yamlutil"
 
 	// Import every production rule package via the shared barrel so
@@ -142,168 +133,6 @@ func printVersion() {
 		}
 	}
 	fmt.Printf("mdsmith %s\n", v)
-}
-
-// runCheck implements the "check" subcommand: lint files.
-func runCheck(args []string) int {
-	fs := flag.NewFlagSet("check", flag.ContinueOnError)
-	var (
-		configPath, format, maxInputSize                              string
-		noColor, quiet, verbose, noGitignore, followSymlinks, explain bool
-	)
-
-	fs.StringVarP(&configPath, "config", "c", "", "Override config file path")
-	fs.StringVarP(&format, "format", "f", "text", "Output format: text, json")
-	fs.BoolVar(&noColor, "no-color", false, "Disable ANSI colors")
-	fs.BoolVarP(&quiet, "quiet", "q", false, "Suppress non-error output")
-	fs.BoolVarP(&verbose, "verbose", "v", false, "Show config, files, and rules on stderr")
-	fs.BoolVar(&noGitignore, "no-gitignore", false, "Disable .gitignore filtering when walking directories")
-	fs.BoolVar(&followSymlinks, "follow-symlinks", false,
-		"Follow symlinks; omitted defers to follow-symlinks config (default skip); "+
-			"=false forces skip over any config opt-in")
-	fs.StringVar(&maxInputSize, "max-input-size", "", "Maximum file size to process (e.g. 2MB, 500KB, 0=unlimited)")
-	fs.BoolVar(&explain, "explain", false, "Attach per-leaf rule provenance to each diagnostic")
-
-	fs.Usage = func() {
-		fmt.Fprintf(os.Stderr, "Usage: mdsmith check [flags] [files...]\n\n"+
-			"Lint Markdown files for style issues.\n\n"+
-			"Files can be paths, directories (walked recursively for *.md), or glob patterns.\n"+
-			"Pass - to read from stdin. With no file arguments, discovers files using the\n"+
-			"files patterns from config (default: **/*.md, **/*.markdown).\n\n"+
-			"Flags:\n")
-		fs.PrintDefaults()
-	}
-
-	if err := fs.Parse(args); err != nil {
-		if code := reportFlagParseErr(err, os.Stderr, "mdsmith: check"); code >= 0 {
-			return code
-		}
-	}
-
-	// --quiet suppresses verbose
-	if quiet {
-		verbose = false
-	}
-
-	walk := walkCLI{
-		noGitignore:    noGitignore,
-		followSymlinks: followSymlinksOverride(fs, followSymlinks),
-	}
-
-	allArgs := fs.Args()
-
-	// Check for explicit stdin argument "-".
-	hasStdin, fileArgs := splitStdinArg(allArgs)
-
-	if hasStdin {
-		return checkStdin(format, noColor, quiet, verbose, configPath, maxInputSize, explain)
-	}
-
-	if len(fileArgs) > 0 {
-		return checkFiles(fileArgs, configPath, format, noColor, quiet, verbose, walk, maxInputSize, explain)
-	}
-
-	// No file args and no stdin: discover files from config.
-	return checkDiscovered(configPath, format, noColor, quiet, verbose, walk, maxInputSize, explain)
-}
-
-// runFix implements the "fix" subcommand: auto-fix lint issues in place.
-func runFix(args []string) int {
-	opts, fileArgs, hasStdin, code := parseFixFlags(args)
-	if code >= 0 {
-		return code
-	}
-	if hasStdin {
-		fmt.Fprintf(os.Stderr, "mdsmith: cannot fix stdin in place\n")
-		return 2
-	}
-	if len(fileArgs) > 0 {
-		return fixFiles(fileArgs, opts)
-	}
-	// No file args: discover files from config.
-	return fixDiscovered(opts)
-}
-
-// parseFixFlags configures the `fix` flag set, parses args, and
-// returns the resolved opts plus positional arguments. The bool
-// `hasStdin` is true when the caller passed `-` as a positional
-// arg. A non-negative `code` means the caller should return that
-// exit code immediately (e.g. --help or a parse error).
-func parseFixFlags(args []string) (fixCLIOpts, []string, bool, int) {
-	fs := flag.NewFlagSet("fix", flag.ContinueOnError)
-	var (
-		configPath, format, maxInputSize                                      string
-		noColor, quiet, verbose, noGitignore, followSymlinks, explain, dryRun bool
-	)
-
-	fs.StringVarP(&configPath, "config", "c", "", "Override config file path")
-	fs.StringVarP(&format, "format", "f", "text", "Output format: text, json")
-	fs.BoolVar(&noColor, "no-color", false, "Disable ANSI colors")
-	fs.BoolVarP(&quiet, "quiet", "q", false, "Suppress non-error output")
-	fs.BoolVarP(&verbose, "verbose", "v", false, "Show config, files, and rules on stderr")
-	fs.BoolVar(&noGitignore, "no-gitignore", false, "Disable .gitignore filtering when walking directories")
-	fs.BoolVar(&followSymlinks, "follow-symlinks", false,
-		"Follow symlinks; omitted defers to follow-symlinks config (default skip); "+
-			"=false forces skip over any config opt-in")
-	fs.StringVar(&maxInputSize, "max-input-size", "", "Maximum file size to process (e.g. 2MB, 500KB, 0=unlimited)")
-	fs.BoolVar(&explain, "explain", false, "Attach per-leaf rule provenance to each remaining diagnostic")
-	fs.BoolVar(&dryRun, "dry-run", false,
-		"Preview which files would change without writing; "+
-			"per-file output lists the rules that would fire and their counts")
-
-	fs.Usage = func() {
-		fmt.Fprintf(os.Stderr, "Usage: mdsmith fix [flags] [files...]\n\n"+
-			"Auto-fix lint issues in Markdown files.\n\n"+
-			"Files can be paths, directories (walked recursively for *.md), or glob patterns.\n"+
-			"Pass - to read from stdin (rejected: files must be writable).\n"+
-			"With no file arguments, discovers files using config patterns.\n\n"+
-			"Flags:\n")
-		fs.PrintDefaults()
-	}
-
-	if err := fs.Parse(args); err != nil {
-		if code := reportFlagParseErr(err, os.Stderr, "mdsmith: fix"); code >= 0 {
-			return fixCLIOpts{}, nil, false, code
-		}
-	}
-
-	// --quiet suppresses verbose
-	if quiet {
-		verbose = false
-	}
-
-	hasStdin, fileArgs := splitStdinArg(fs.Args())
-
-	return fixCLIOpts{
-		configPath: configPath,
-		format:     format,
-		noColor:    noColor,
-		quiet:      quiet,
-		verbose:    verbose,
-		walk: walkCLI{
-			noGitignore:    noGitignore,
-			followSymlinks: followSymlinksOverride(fs, followSymlinks),
-		},
-		maxInputSize: maxInputSize,
-		explain:      explain,
-		dryRun:       dryRun,
-	}, fileArgs, hasStdin, -1
-}
-
-// fixCLIOpts bundles the runtime knobs threaded through the fix
-// command path. Grouped because runFix splits between explicit-file
-// and config-discovery entry points and the same nine values flow
-// to both.
-type fixCLIOpts struct {
-	configPath   string
-	format       string
-	noColor      bool
-	quiet        bool
-	verbose      bool
-	walk         walkCLI
-	maxInputSize string
-	explain      bool
-	dryRun       bool
 }
 
 // runQuery implements the "query" subcommand: select files by CUE
@@ -499,171 +328,6 @@ func runInit(args []string) int {
 	return 0
 }
 
-// printDryRunPreview writes one line per file that would change.
-// Files where some diagnostics would be resolved get a
-// "would fix N violations (MDS001 ×2, MDS006)" line; files whose
-// bytes would change without any diagnostic count decreasing get a
-// "would update generated content" line so a dry-run still surfaces
-// directive regeneration. No-op when the preview is empty. Write
-// errors on the destination are intentionally ignored: the preview
-// is supplemental context, not the primary signal (exit code and
-// diagnostic output drive CI), and a half-written preview is not
-// worth halting the run for.
-func printDryRunPreview(w io.Writer, fixResult *fixpkg.Result) {
-	for _, wf := range fixResult.WouldFixFiles {
-		if wf.Count == 0 {
-			_, _ = fmt.Fprintf(w, "%s: would update generated content\n", wf.Path)
-			continue
-		}
-		_, _ = fmt.Fprintf(w, "%s: would fix %s\n", wf.Path, formatWouldFixSummary(wf))
-	}
-}
-
-// formatWouldFixSummary renders "N violations (MDS001 ×2, MDS006)"
-// for one file's would-fix entry. Single-count rules are printed
-// without a multiplier so the common case stays compact.
-func formatWouldFixSummary(wf fixpkg.WouldFixFile) string {
-	parts := make([]string, 0, len(wf.Rules))
-	for _, r := range wf.Rules {
-		if r.Count == 1 {
-			parts = append(parts, r.RuleID)
-			continue
-		}
-		parts = append(parts, fmt.Sprintf("%s ×%d", r.RuleID, r.Count))
-	}
-	noun := "violations"
-	if wf.Count == 1 {
-		noun = "violation"
-	}
-	if len(parts) == 0 {
-		return fmt.Sprintf("%d %s", wf.Count, noun)
-	}
-	return fmt.Sprintf("%d %s (%s)", wf.Count, noun, strings.Join(parts, ", "))
-}
-
-// dryRunJSONFile is the per-file JSON shape produced by
-// `mdsmith fix --dry-run --format json`. One record per file
-// whose bytes or diagnostic counts would change — fixable
-// violations, remaining unfixable diagnostics, or
-// generated-section regeneration (would_fix=0, rules=[]).
-type dryRunJSONFile struct {
-	Path        string     `json:"path"`
-	WouldFix    int        `json:"would_fix"`
-	Rules       []string   `json:"rules"`
-	Diagnostics []jsonDiag `json:"diagnostics"`
-}
-
-// jsonDiag mirrors the diagnostic shape JSONFormatter writes today
-// so dry-run JSON records carry the same per-diagnostic fields,
-// including the optional source-context and explanation trailers.
-type jsonDiag struct {
-	File            string       `json:"file"`
-	Line            int          `json:"line"`
-	Column          int          `json:"column"`
-	Rule            string       `json:"rule"`
-	Name            string       `json:"name"`
-	Severity        string       `json:"severity"`
-	Message         string       `json:"message"`
-	SourceLines     []string     `json:"source_lines,omitempty"`
-	SourceStartLine int          `json:"source_start_line,omitempty"`
-	Explanation     *jsonDiagExp `json:"explanation,omitempty"`
-}
-
-// jsonDiagExp is the explanation trailer shape, matching
-// output.jsonExplanation so dry-run JSON is schema-compatible with
-// check --format json.
-type jsonDiagExp struct {
-	Rule   string         `json:"rule"`
-	Leaves []jsonDiagLeaf `json:"leaves"`
-}
-
-// jsonDiagLeaf is one leaf entry inside a jsonDiagExp.
-type jsonDiagLeaf struct {
-	Path   string `json:"path"`
-	Value  any    `json:"value"`
-	Source string `json:"source"`
-}
-
-// diagsToJSONDiags converts lint diagnostics to the JSON shape used by
-// dry-run output.
-func diagsToJSONDiags(diags []lint.Diagnostic) []jsonDiag {
-	jsonDiags := make([]jsonDiag, 0, len(diags))
-	for _, d := range diags {
-		var exp *jsonDiagExp
-		if d.Explanation != nil {
-			leaves := make([]jsonDiagLeaf, 0, len(d.Explanation.Leaves))
-			for _, l := range d.Explanation.Leaves {
-				leaves = append(leaves, jsonDiagLeaf{Path: l.Path, Value: l.Value, Source: l.Source})
-			}
-			exp = &jsonDiagExp{Rule: d.Explanation.Rule, Leaves: leaves}
-		}
-		jsonDiags = append(jsonDiags, jsonDiag{
-			File:            d.File,
-			Line:            d.Line,
-			Column:          d.Column,
-			Rule:            d.RuleID,
-			Name:            d.RuleName,
-			Severity:        string(d.Severity),
-			Message:         d.Message,
-			SourceLines:     d.SourceLines,
-			SourceStartLine: d.SourceStartLine,
-			Explanation:     exp,
-		})
-	}
-	return jsonDiags
-}
-
-// writeDryRunJSON emits the per-file dry-run JSON payload on w as a
-// JSON array of dryRunJSONFile records. Records are emitted for every
-// file in WouldFixFiles plus any file that has remaining diagnostics
-// not already covered. Callers route w to stderr to match the
-// existing lint-output contract documented in
-// docs/reference/cli.md. Returns a non-zero exit code on write error.
-func writeDryRunJSON(w io.Writer, fixResult *fixpkg.Result) int {
-	diagsByFile := make(map[string][]lint.Diagnostic)
-	for _, d := range fixResult.Diagnostics {
-		diagsByFile[d.File] = append(diagsByFile[d.File], d)
-	}
-	seen := make(map[string]struct{}, len(fixResult.WouldFixFiles))
-	records := make([]dryRunJSONFile, 0, len(fixResult.WouldFixFiles))
-	for _, wf := range fixResult.WouldFixFiles {
-		seen[wf.Path] = struct{}{}
-		rules := make([]string, 0, len(wf.Rules))
-		for _, r := range wf.Rules {
-			rules = append(rules, r.RuleID)
-		}
-		records = append(records, dryRunJSONFile{
-			Path:        wf.Path,
-			WouldFix:    wf.Count,
-			Rules:       rules,
-			Diagnostics: diagsToJSONDiags(diagsByFile[wf.Path]),
-		})
-	}
-	// Include files with only unfixable diagnostics (not in WouldFixFiles).
-	unfixable := make([]string, 0, len(diagsByFile))
-	for path := range diagsByFile {
-		if _, ok := seen[path]; !ok {
-			unfixable = append(unfixable, path)
-		}
-	}
-	sort.Strings(unfixable)
-	for _, path := range unfixable {
-		records = append(records, dryRunJSONFile{
-			Path:        path,
-			WouldFix:    0,
-			Rules:       []string{},
-			Diagnostics: diagsToJSONDiags(diagsByFile[path]),
-		})
-	}
-	enc := json.NewEncoder(w)
-	enc.SetIndent("", "  ")
-	if err := enc.Encode(records); err != nil {
-		fmt.Fprintf(os.Stderr, "mdsmith: error writing dry-run output: %v\n", err)
-		return 2
-	}
-	return 0
-}
-
 // formatDiagnosticsTo writes diagnostics to w using the specified format.
 // Returns a non-zero exit code on write error, or 0 on success.
 func formatDiagnosticsTo(w io.Writer, diags []lint.Diagnostic, format string, noColor bool) int {
@@ -745,157 +409,6 @@ func printRunStatsTo(w io.Writer, format string, quiet bool, stats runStats) {
 		stats.Failures,
 		stats.Unfixed,
 	)
-}
-
-// checkFiles lints the given file paths and returns the appropriate exit code.
-func checkFiles(
-	fileArgs []string, configPath, format string,
-	noColor, quiet, verbose bool, walk walkCLI,
-	maxInputSize string, explain bool,
-) int {
-	cfg, cfgPath, logger, files, maxBytes, code := loadAndResolve(
-		fileArgs, configPath, verbose, walk, maxInputSize,
-	)
-	if code >= 0 {
-		return code
-	}
-
-	runner := &engine.Runner{
-		Config:           cfg,
-		Rules:            rule.All(),
-		StripFrontMatter: frontMatterEnabled(cfg),
-		Logger:           logger,
-		RootDir:          rootDirFromConfig(cfgPath),
-		MaxInputBytes:    maxBytes,
-		Explain:          explain,
-		ConfigPath:       cfgPath,
-	}
-	result := runner.Run(files)
-	printErrors(result.Errors)
-
-	if !quiet && len(result.Diagnostics) > 0 {
-		if code := formatDiagnostics(result.Diagnostics, format, noColor); code != 0 {
-			return code
-		}
-	}
-	printRunStats(format, quiet, runStats{
-		Checked:  result.FilesChecked,
-		Fixed:    0,
-		Failures: len(result.Diagnostics),
-		Unfixed:  len(result.Diagnostics),
-	})
-	logger.Printf("checked %d files, %d issues found", result.FilesChecked, len(result.Diagnostics))
-
-	if len(result.Errors) > 0 && len(result.Diagnostics) == 0 {
-		return 2
-	}
-	if len(result.Diagnostics) > 0 {
-		return 1
-	}
-	return 0
-}
-
-// fixFiles fixes lint issues in the given file paths.
-func fixFiles(fileArgs []string, opts fixCLIOpts) int {
-	cfg, cfgPath, logger, files, maxBytes, code := loadAndResolve(
-		fileArgs, opts.configPath, opts.verbose, opts.walk, opts.maxInputSize,
-	)
-	if code >= 0 {
-		return code
-	}
-
-	fixer := &fixpkg.Fixer{
-		Config:           cfg,
-		Rules:            rule.All(),
-		StripFrontMatter: frontMatterEnabled(cfg),
-		Logger:           logger,
-		RootDir:          rootDirFromConfig(cfgPath),
-		MaxInputBytes:    maxBytes,
-		Explain:          opts.explain,
-		DryRun:           opts.dryRun,
-	}
-	fixResult := fixer.Fix(files)
-	return reportFixResult(opts, fixResult, logger)
-}
-
-// readStdinLimited reads stdin with an optional size limit.
-// When maxBytes <= 0 no limit is applied.
-func readStdinLimited(maxBytes int64) ([]byte, error) {
-	// Treat MaxInt64 as unlimited to avoid overflow in the +1 sentinel.
-	if maxBytes > 0 && maxBytes < math.MaxInt64 {
-		data, err := io.ReadAll(io.LimitReader(os.Stdin, maxBytes+1))
-		if err != nil {
-			return nil, err
-		}
-		if int64(len(data)) > maxBytes {
-			return nil, fmt.Errorf(
-				"reading \"<stdin>\": file too large "+
-					"(%d bytes, max %d)", int64(len(data)), maxBytes)
-		}
-		return data, nil
-	}
-	return io.ReadAll(os.Stdin)
-}
-
-// checkStdin reads from stdin, lints the content, and returns the appropriate
-// exit code. Uses runner.RunSource to ensure Configurable settings are applied.
-func checkStdin(format string, noColor, quiet, verbose bool, configPath, maxInputSize string, explain bool) int {
-	logger := &vlog.Logger{Enabled: verbose, W: os.Stderr}
-
-	cfg, cfgPath, err := loadConfig(configPath)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "mdsmith: %v\n", err)
-		return 2
-	}
-	if cfgPath != "" {
-		logger.Printf("config: %s", cfgPath)
-	}
-
-	maxBytes, err := resolveMaxInputBytes(cfg, maxInputSize)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "mdsmith: %v\n", err)
-		return 2
-	}
-
-	source, err := readStdinLimited(maxBytes)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "mdsmith: %v\n", err)
-		return 2
-	}
-
-	runner := &engine.Runner{
-		Config:           cfg,
-		Rules:            rule.All(),
-		StripFrontMatter: frontMatterEnabled(cfg),
-		Logger:           logger,
-		RootDir:          rootDirFromConfig(cfgPath),
-		MaxInputBytes:    maxBytes,
-		Explain:          explain,
-		ConfigPath:       cfgPath,
-	}
-	result := runner.RunSource("<stdin>", source)
-	printErrors(result.Errors)
-
-	if !quiet && len(result.Diagnostics) > 0 {
-		if code := formatDiagnostics(result.Diagnostics, format, noColor); code != 0 {
-			return code
-		}
-	}
-	printRunStats(format, quiet, runStats{
-		Checked:  result.FilesChecked,
-		Fixed:    0,
-		Failures: len(result.Diagnostics),
-		Unfixed:  len(result.Diagnostics),
-	})
-	logger.Printf("checked %d files, %d issues found", result.FilesChecked, len(result.Diagnostics))
-
-	if len(result.Errors) > 0 && len(result.Diagnostics) == 0 {
-		return 2
-	}
-	if len(result.Diagnostics) > 0 {
-		return 1
-	}
-	return 0
 }
 
 // loadAndResolve loads config, resolves file paths, and parses the max
@@ -984,136 +497,6 @@ func discoverFiles(
 		return nil, "", nil, nil, 0
 	}
 	return cfg, cfgPath, logger, files, -1
-}
-
-// checkDiscovered loads config, discovers files from config patterns,
-// and lints them. Returns the appropriate exit code.
-func checkDiscovered(
-	configPath, format string,
-	noColor, quiet, verbose bool, walk walkCLI,
-	maxInputSize string, explain bool,
-) int {
-	cfg, cfgPath, logger, files, code := discoverFiles(configPath, verbose, walk)
-	if code >= 0 {
-		return code
-	}
-
-	maxBytes, err := resolveMaxInputBytes(cfg, maxInputSize)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "mdsmith: %v\n", err)
-		return 2
-	}
-
-	runner := &engine.Runner{
-		Config:           cfg,
-		Rules:            rule.All(),
-		StripFrontMatter: frontMatterEnabled(cfg),
-		Logger:           logger,
-		RootDir:          rootDirFromConfig(cfgPath),
-		MaxInputBytes:    maxBytes,
-		Explain:          explain,
-		ConfigPath:       cfgPath,
-	}
-	result := runner.Run(files)
-	printErrors(result.Errors)
-
-	if !quiet && len(result.Diagnostics) > 0 {
-		if code := formatDiagnostics(result.Diagnostics, format, noColor); code != 0 {
-			return code
-		}
-	}
-	printRunStats(format, quiet, runStats{
-		Checked:  result.FilesChecked,
-		Fixed:    0,
-		Failures: len(result.Diagnostics),
-		Unfixed:  len(result.Diagnostics),
-	})
-	logger.Printf("checked %d files, %d issues found", result.FilesChecked, len(result.Diagnostics))
-
-	if len(result.Errors) > 0 && len(result.Diagnostics) == 0 {
-		return 2
-	}
-	if len(result.Diagnostics) > 0 {
-		return 1
-	}
-	return 0
-}
-
-// fixDiscovered loads config, discovers files from config patterns,
-// and fixes them. Returns the appropriate exit code.
-func fixDiscovered(opts fixCLIOpts) int {
-	cfg, cfgPath, logger, files, code := discoverFiles(opts.configPath, opts.verbose, opts.walk)
-	if code >= 0 {
-		return code
-	}
-
-	maxBytes, err := resolveMaxInputBytes(cfg, opts.maxInputSize)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "mdsmith: %v\n", err)
-		return 2
-	}
-
-	fixer := &fixpkg.Fixer{
-		Config:           cfg,
-		Rules:            rule.All(),
-		StripFrontMatter: frontMatterEnabled(cfg),
-		Logger:           logger,
-		RootDir:          rootDirFromConfig(cfgPath),
-		MaxInputBytes:    maxBytes,
-		Explain:          opts.explain,
-		DryRun:           opts.dryRun,
-	}
-	fixResult := fixer.Fix(files)
-	return reportFixResult(opts, fixResult, logger)
-}
-
-// reportFixResult writes the fix run's output and computes the exit
-// code. Shared by fixFiles and fixDiscovered so the dry-run preview,
-// stats summary, and exit-code logic stay in one place.
-func reportFixResult(opts fixCLIOpts, fixResult *fixpkg.Result, logger *vlog.Logger) int {
-	return reportFixResultTo(opts, fixResult, logger, os.Stderr)
-}
-
-// reportFixResultTo is the injectable form of reportFixResult. Tests
-// pass an alternate stderr writer to exercise the write-error
-// branches without leaking to the real stderr.
-func reportFixResultTo(opts fixCLIOpts, fixResult *fixpkg.Result, logger *vlog.Logger, stderrW io.Writer) int {
-	printErrorsTo(stderrW, fixResult.Errors)
-
-	if opts.dryRun && opts.format == "json" && !opts.quiet {
-		// Match `check --format json` and `fix --format json`: lint
-		// output goes to stderr (see docs/reference/cli.md "Output").
-		if code := writeDryRunJSON(stderrW, fixResult); code != 0 {
-			return code
-		}
-	} else {
-		if opts.dryRun && !opts.quiet {
-			printDryRunPreview(stderrW, fixResult)
-		}
-		if !opts.quiet && len(fixResult.Diagnostics) > 0 {
-			if code := formatDiagnosticsTo(stderrW, fixResult.Diagnostics, opts.format, opts.noColor); code != 0 {
-				return code
-			}
-		}
-	}
-
-	printRunStatsTo(stderrW, opts.format, opts.quiet, runStats{
-		Checked:  fixResult.FilesChecked,
-		Fixed:    len(fixResult.Modified),
-		Failures: fixResult.Failures,
-		Unfixed:  len(fixResult.Diagnostics),
-		WouldFix: fixResult.WouldFix,
-		DryRun:   opts.dryRun,
-	})
-	logger.Printf("checked %d files, %d issues found", fixResult.FilesChecked, len(fixResult.Diagnostics))
-
-	if len(fixResult.Errors) > 0 && len(fixResult.Diagnostics) == 0 {
-		return 2
-	}
-	if len(fixResult.Diagnostics) > 0 {
-		return 1
-	}
-	return 0
 }
 
 // walkCLI bundles the CLI flags that affect how files are
@@ -1279,264 +662,4 @@ func printDeprecations(cfg *config.Config) {
 		fmt.Fprintf(os.Stderr, "mdsmith: deprecated: %s\n", msg)
 	}
 	cfg.Deprecations = nil
-}
-
-const helpUsageText = `Usage: mdsmith help <topic>
-
-Topics:
-  rule [id|name]        Show rule documentation
-  metrics [id|name]     Show metric documentation
-  kinds                 Show concept page for file kinds
-  kinds-cli             Summarize the 'kinds' subcommand surface
-  placeholder-grammar   Show placeholder vocabulary reference
-  patterns              Show maintainability patterns across rules
-`
-
-// runHelp implements the "help" subcommand.
-func runHelp(args []string) int {
-	if len(args) == 0 {
-		fmt.Fprint(os.Stderr, helpUsageText)
-		return 0
-	}
-
-	switch args[0] {
-	case "rule":
-		return runHelpRule(args[1:])
-	case "metrics":
-		return runHelpMetrics(args[1:])
-	case "kinds":
-		return runHelpKinds()
-	case "kinds-cli":
-		return runHelpKindsCLI()
-	case "placeholder-grammar":
-		return runHelpConcept("placeholder-grammar")
-	case "patterns":
-		return runHelpPatterns(args[1:])
-	default:
-		fmt.Fprintf(os.Stderr, "mdsmith: help: unknown topic %q\n", args[0])
-		return 2
-	}
-}
-
-// listRulesForHelp is the ruledocs.ListRules dependency, indirected through
-// a package var so tests can substitute a fault-injecting lister and exercise
-// the error-handling branches in the help subcommands.
-var listRulesForHelp = ruledocs.ListRules
-
-type patternRec struct {
-	ID            string `json:"id"`
-	Name          string `json:"name"`
-	Signal        string `json:"signal"`
-	Fix           string `json:"fix"`
-	ForDiagnostic bool   `json:"for-diagnostic"`
-}
-
-func runHelpPatterns(args []string) int {
-	format, code, ok := parsePatternsFormat(args)
-	if !ok {
-		return code
-	}
-	rules, err := listRulesForHelp()
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "mdsmith: %v\n", err)
-		return 2
-	}
-	items := make([]patternRec, 0)
-	for _, r := range rules {
-		if r.Maintainability == nil {
-			continue
-		}
-		items = append(items, patternRec{
-			ID:            r.ID,
-			Name:          r.Name,
-			Signal:        r.Maintainability.Signal,
-			Fix:           r.Maintainability.Fix,
-			ForDiagnostic: r.Maintainability.ForDiagnostic,
-		})
-	}
-	if format == "json" {
-		enc := json.NewEncoder(os.Stdout)
-		enc.SetIndent("", "  ")
-		if err := enc.Encode(items); err != nil {
-			fmt.Fprintf(os.Stderr, "mdsmith: writing json: %v\n", err)
-			return 2
-		}
-		return 0
-	}
-	for _, it := range items {
-		fmt.Printf("%s %s\n  signal: %s\n  fix: %s\n  for-diagnostic: %t\n\n",
-			it.ID, it.Name, it.Signal, it.Fix, it.ForDiagnostic)
-	}
-	return 0
-}
-
-// parsePatternsFormat extracts the --format value from runHelpPatterns args.
-// Returns (format, exitCode, ok); when ok is false the caller should return
-// exitCode immediately. Accepts: no args (text), `-f|--format <text|json>`.
-func parsePatternsFormat(args []string) (string, int, bool) {
-	if len(args) == 0 {
-		return "text", 0, true
-	}
-	if args[0] != "-f" && args[0] != "--format" {
-		fmt.Fprintf(os.Stderr, "mdsmith: help patterns: unexpected argument %q\n", args[0])
-		return "", 2, false
-	}
-	if len(args) < 2 {
-		fmt.Fprintf(os.Stderr, "mdsmith: help patterns: %s requires a value (text or json)\n", args[0])
-		return "", 2, false
-	}
-	if len(args) > 2 {
-		fmt.Fprintf(os.Stderr, "mdsmith: help patterns: unexpected trailing argument %q\n", args[2])
-		return "", 2, false
-	}
-	format := args[1]
-	if format != "text" && format != "json" {
-		fmt.Fprintf(os.Stderr, "mdsmith: help patterns: unknown format %q (valid: text, json)\n", format)
-		return "", 2, false
-	}
-	return format, 0, true
-}
-
-const helpKindsText = `File Kinds
-
-A kind is a named bundle of rule settings that can be applied to a set of
-files. Kinds let you share per-rule tuning across files that serve the same
-purpose (schema, template, fragment, prompt, …) without repeating overrides.
-
-DECLARATION
-
-Declare kinds under the kinds: key. The body has the same shape as an
-override entry (rules:, categories:) — minus glob:, since files are bound
-to kinds separately:
-
-  kinds:
-    plan:
-      rules:
-        required-structure:
-          schema: plan/proto.md
-        paragraph-readability: false
-    proto:
-      rules:
-        paragraph-readability: false
-        first-line-heading: false
-
-Kind names are project-chosen. mdsmith ships no built-in kinds.
-
-ASSIGNMENT
-
-A file's effective kind list is built from two sources, concatenated in
-this order:
-
-  1. Front-matter kinds: field (YAML list).
-  2. Matching entries in kind-assignment: (config order; each entry's kinds
-     in the order listed).
-
-Duplicate names are dropped after their first occurrence. Referencing an
-undeclared kind is a config error.
-
-  kind-assignment:
-    - glob: ["plan/[0-9]*_*.md"]
-      kinds: [plan]
-    - glob: ["**/proto.md"]
-      kinds: [proto]
-
-MERGE ORDER
-
-Kinds apply after top-level rules and before glob overrides:
-
-  top-level rules → kinds (effective-list order) → glob overrides
-
-Within kinds, the later kind in the effective list replaces the earlier
-kind's entire rule config for that rule — no deep-merge, same as overrides.
-A file's own glob overrides apply last and take highest precedence.
-
-COMPOSABILITY
-
-Rules never reference kind names. New kinds cannot regress existing behavior.
-`
-
-// runHelpKinds prints the kinds concept page.
-func runHelpKinds() int {
-	fmt.Print(helpKindsText)
-	return 0
-}
-
-const helpKindsCLIText = `Kinds Subcommand
-
-mdsmith kinds <subcommand> [args]
-
-Subcommands:
-  list                  Print declared kinds with their merged bodies.
-  show <name>           Print one kind's merged body.
-  path <name>           Print the resolved schema path of the kind's
-                        required-structure rule, if any.
-  resolve <file>        Print the resolved kind list and merged rule
-                        config for a file, with per-leaf provenance.
-  why <file> <rule>     Print the full merge chain for one rule on
-                        one file: every applicable layer, including
-                        no-ops, with the value at each step.
-
-Each subcommand accepts --json for stable structured output. The
-schema is documented in docs/reference/cli.md.
-
-Provenance layers are: 'default' (top-level rules: + built-ins),
-'kinds.<name>' (one per kind in the effective list), and
-'overrides[<i>]' (one per matching glob override entry).
-
-See also: 'mdsmith check --explain' / 'mdsmith fix --explain' to
-attach the same provenance trailer to each diagnostic.
-`
-
-// runHelpKindsCLI prints the kinds-cli help topic.
-func runHelpKindsCLI() int {
-	fmt.Print(helpKindsCLIText)
-	return 0
-}
-
-// runHelpConcept prints the named concept page.
-func runHelpConcept(name string) int {
-	content, err := concepts.Lookup(name)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "mdsmith: %v\n", err)
-		return 2
-	}
-	fmt.Print(content)
-	return 0
-}
-
-// runHelpRule implements "help rule [id|name]".
-func runHelpRule(args []string) int {
-	if len(args) == 0 {
-		return listAllRules()
-	}
-	return showRule(args[0])
-}
-
-func listAllRules() int {
-	rules, err := ruledocs.ListRules()
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "mdsmith: %v\n", err)
-		return 2
-	}
-
-	for _, r := range rules {
-		fmt.Printf("%-6s %-40s %-10s %s\n", r.ID, r.Name, r.Status, r.Description)
-	}
-	return 0
-}
-
-func showRule(query string) int {
-	info, err := ruledocs.LookupRuleInfo(query)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "mdsmith: %v\n", err)
-		return 2
-	}
-	content := ruledocs.StripFrontMatter(info.Content)
-	if m := info.Maintainability; m != nil {
-		content += "\n\n## Maintainability pattern\n\n"
-		content += fmt.Sprintf("- Signal: %s\n- Fix: %s\n- For diagnostic: %t\n",
-			m.Signal, m.Fix, m.ForDiagnostic)
-	}
-	fmt.Print(content)
-	return 0
 }
